Make last change work for unique constraints
DAT-19267
